Interresting commit that make us switch to version 0.71 !
Changes affects a lot of files :

- Cleaned the new dependences API to allow a sound id naming and renaming scheme,
and placed all the sound API in vm/deps.c. Now the objects are not loaded multiple
times if they are already loaded, dependences cannot be unloaded if one of their
root file is still loaded, dependence can change id if one of their root file is
unloaded and another root file is still loaded, so that id prefix keeps beeing
sound.

- Created search.c, that now contains a reusable API for searching in a pool of
objects, by supplying hashes tables to work on (useful for tracer, but not only)

- Added 5 functions to the Hash API (libelfsh/libhash.c), especially hash_merge(),
hash_unmerge(), and hash_size().

- Added the abstract "force" prefix (see help)

- Cleaned the code a little bit everywhere to use more hash_get_keys() and less
nested for loops