- Added a type for elements inside vectors and hash tables. Now aspect_register _vector and hash_init takes one more parameter. See libaspect.h for all possible types. This allow to read/write framework hash/vectors elements inside scripting :)
- Changed librevm parser in order to use a hash table of parsing handlers. That
allow to register new syntax for the scripting langage more easily.

- Added a flag in the revmL1_t object in order to distinguish of the children
are L1 or L2. In case L1 is a child of L1, that allow to do 1.L1.L1.L1......L1.L1.L2
indexations without size limit in the scripting. That will be useful when
we will be able to register new objects in the framework and bind them
transparently in the scripting langage.

- Added a missing check in vm_lookup_param() in grammar.c to avoid wrong
success message when using the 'set' command wrongly.