First of all, sorry for the delay. I've been quite busy these days.
Ok, I tried to create the operand handlers for SPARC and put them into their own vector (called "operands-sparc").
I tested it here and apparently everything is working fine. Plz kick me if smt goes wrong.


A few comments, though:

- The handlers have their own vector (separated from IA-32) because they work on different types. The IA-32 types
are the ones in the enum e_asm_operand_types (libasm-i386.h) and the SPARC types are the ones in a unnamed enum
in libasm-sparc.h (the values are prefixed by ASM_SP_OTYPE_). While the architectures work on so different types
and have such a different number of types, making them use the same vector would be just insane and senseless.

- I know that an unified type system is the ideal, but the architectures have so different types and the code is
so type dependent (the written output depend on the type, for instance), that the efforts to unify the type system
would be fairly big and therefore had to be delayed.

- All the handlers do is set the 'type' member of the asm_operand structure. There's no code setting the 'base_reg'
or 'imm' member fields, for instance. Why? Because the code that does that depends on the format of the instruction
and on the instruction itself. So in order to do that, the code inside the operand handlers would need to have big
conditionals testing cases for all the different instructions and etc. Again, alternatives to this approach can be
found, but may require lots of effort and major modifications in the library architecture (and I honestly don't
think it's worth it).

- If someone wants to add more complex code to the handlers, eg. code that investigate the values of the operands,
they'll probably have to use the sparc_convert_* functions, which have been adapted to work from inside the handlers
(they don't need an asm_processor * parameter anymore).


So, apart of these problems, the main goal of this patch has been accomplished: refactoring the source so that we can
easily add code to be run everytime an operand of a specific type is found (all we have to do is open the related file
inside libasm/src/arch/sparc/operand_handlers/ and code).

Cheers,

    Strauss



Ps.: Below, a command and a small script that helped me to do the most repetitive part of the job and that, with some adaption,
will probably help me in the future as well (you don't have to look at them, I'm just posting them here for backup reasons):

grep -h sparc_convert src/arch/sparc/handlers/* | sort | uniq



#!/bin/bash
#libasm-replace.sh

for file in `find handlers/ -type f`
do
  sed s%'sparc_convert_format4(&opcode4 buf);'%'sparc_convert_format4(\&opcode4, buf);'%g $file > $file.new
  mv $file.new $file
done