The first stable feature of ERESI expressions : printing
declare.esh was updated to reflect this.

Also:

- lts.c was removed as it was a double with convert.c

- Typed expression instanciation is compatible with what we
need for program transformation. We support variable declaration
both :

struct0 mystruct0 = (field0:42)

and

struct0 mystruct0b = struct0(field0:42)

In case a type is given (second case), type checking is done to make sure
we type well. This feature sounds useless, but typename(...) is what we
need for implementing the "match" and this simple feature will help a lot
discarding all matching candidate that has a different root type. Additionally,
lots of error checking is done so any error in the expression declaration will
forbid to create the variable.

- Added a hash table exprs_hash that does $expr_name -> expr* and automatically
manage that hash table when creating expressions.

- Improved printing of revmobj_t (added support for raw fields, arrays, and
added colors)

- Switched to version 0.79b1

TODO:

- Code the comparison/set between expressions. Its already almost done but
I lost a lot of time today with pretty printing the expressions ..