- Nice improvement of libmjollnir :
* No duplicate code anymore for storing/loading on/from disk. I
put all the common code in : libmjollnir/src/ondisk.c. Now the
saving/loading code can be used for -ANY- kind of data that is
inside containers.

* Switched mjrlink_t *output and *input lists into list_t of mjrlink_t
so that we can use the ERESI "foreach" command on those lists.

* Now the ->type field of containers refers to an -ERESI TYPE- and
not just block, func, or container type. That allows us now to put
ANY kind of data in containers, and even reuse the routines for
storing and loading from disk with very minimal effort !

* Removed the MJR_LINK_DELETE type and mjr_link_cleanup()
system and did the equivalent code in the good way to avoid
this wacky system.

* Put all structures of libmjollnir into libmjollnir-objects.h and
removed useless headers (libmjollnir had too many headers)

* Adapted graph, flowjack, and inspect commands to use the new API

* Widely factored code everywhere

-> All code was tested on /bin/ls on linux/x86 (analyse, save, reload)


- Various automatic cleaning in the whole project (sorry if the the
modified files list is very big)

- Switched to version 0.8a6