Several new commands and bug patching in this 0.8a12. Note that there will be a 0.8a13 because this commit still has a bug, but I had to do it for making pouik capable to test "redir" in the kernel.
So the changes are:

- Implemented multiple new commands : reflect, define, undef. Reflect allow
to create a list of instruction on-the-fly to get iterated using "foreach",
define allow to define a name to a constant, undef allow to cancel it.

- Patched a newly introduced bug in the CFLOW technique when done in runtime.
That is now working again, and that will make pouik capable to use redir
on the kernel.

- Removed hashent_t and replaced all of them with listent_t which was the
same structure anyway. That will also facilitate transactions between
hash table and lists.

- Added one more testsuite file foreach-linearity.esh that I had forgotten

Enjoy

-jfv