The client can communicate with VMWare without crashing, that's a good sign.  For now, we deal with a 32b VM, I haven't looked at how gdb decides to use 32 or 64b.
I redefined some types, like uint8_t to uint8, it's less boring. I
defined a type called "ureg32" for 32b registers. I'm gonna do the
same for 64b registers and a type "ureg" which will take the right
size (size of the host). Macros to take the LOWORD, HIWORD, etc. have
been added, too.  I had to declare a special structure for the
register gdb uses. It is mixing the register declarations, for example
edx comes before ebx :S. Since in the gdb protocol we have the
register number, it was mandatory to do smth like that :/. For
example, register 8 is ip, so we can reach it by doing
(&begining_structure) + reg_number * sizeof(ureg32).

Split the header in 2 parts. One part is for the internals, the other
one is for the "wide use".  I'm probably gonna split the gdbwrapper.c
file too. We have 2 major parts in this file: the "internals", like
string parsing, message creation, atoh functions, checksum
calculation... and the other part dealing with the network.