- Summary: Working graphs on the linux kernel
--NOTE THAT THIS IS A BACKUP COMMIT AND THERE ARE 
3 REMAINING FIXME IN THE CODE--

- Modified libmjollnir in order to do "on-demand"
analysis if we pass a parameter to "analyse" command.
We never analyse an entire section now, but we read
the section's data little by little while constructing
the graph. Removed useless code in libmjollnir for
old global analysis.

- Moved goto hash table to the context structure in 
libmjollnir. Named dynamically "blocks" / "functions"
and "links" tables to $filename_hashname so that
we can -really- analyse multiple files at the same
time.

- Better system for cleaning an existing analysis. We
dont have an "analysed" flag, but we lookup if a
given address was already analysed.

- Fixed bug in libmjollnir that was not syncing the
symbol table after injecting a symbol. This was making
the analysis of the kernel to fail. See FIXME.

- We dont have anymore to use analyse + graph, but 
directly graph will work.

- Many fixes in libasm for 16b and 64b support. Also
included many fixes for starting to use libasm-arm.

- Better handling of the asm_processor. Do not reinitialize
a new proc when not necessary. Added librevm/api/asmproc.c 
for modular handling, added a parameter to mjr_setup_proc()
to reuse the processor if possible. Renamed world.proc to
world.proc_ia32 since it was badly named.

- Allow dynamic names for hash tables and lists. Adapted
commands "empty", "tables" and "lists" to be able to 
handle this. Modified evarista for using those new dynamic
names. A new lookup function: revm_lookup_key() do that
job on a modular basis.

- Now really make use of the mjr.cfg_depth configuration
variable. Changing it now has an effect. If you graph code
and then decide to change the analysis depth, you need
to strip the previous analysis using a new command
"astrip" and graph again.

- Remove the LD_PRELOAD variable before doing any external
execution in e2dbg. This allows us to generate graphs
from e2dbg. 

- Now switch the current file when reaching a breakpoint
in e2dbg.

-----------------------------------

The 3 remaining FIXME are:

- There is an introduced bug in the displaying of 
disassembling of 64bits architecture (e.g. sparc)
that I am about to fix.

- There is some bug remaining in evarista that makes it segfault
at some point and I still have to discover why. It seems to
be linked with the dynamic names for hash tables.

- Because we sync the symbol table at every injected symbol,
we have to sort it again at each symbol injection. This
makes libmjollnir slower, especially on the kernel which
has 32000 symbols. This can be fixed by implementing a better
system for syncing the sorted symbol table (stored in 
symsect->altdata). For example we could use a btree (as
provided by libaspect-btree.h) so that (sorted) insertion and
lookup takes a logarithmic time.