kedbg & libgdbwrap: Added NoACK mode support, some other bugfixes and a draft of the ARM regs context struct.
Finally, the problem with ACK's missing and/or infinite send() loop is solved.
There still some issues in the way that gdbwrap handles packets in this new mode
but we can safely move on. Also, there is something strange in gdbwrap_atoh()
which is just triggered after we do some stepping. A crazy workaround was added,
and this is not blocking us anymore but still needing some love.